+++ ui-ngx/src/app/modules/home/pages/LwM2M/security/Client Security Config/clientSecurityConfig.component.ts
@@ -3,7 +3,8 @@ import { AfterViewInit, Component, OnInit, ViewChild } from "@angular/core";
 import { MatDialog } from "@angular/material/dialog";
 import { MatPaginator } from "@angular/material/paginator";
 import { MatTableDataSource } from "@angular/material/table";
-import {newConfigDialog} from "./newConfig-dialog.component"
+import { LwService } from "../../Lw-service";
+import { newConfigDialog } from "./newConfig-dialog.component"
 
 @Component({
     selector: 'Lw-clientSecurityConfig',
@@ -14,24 +15,24 @@ import {newConfigDialog} from "./newConfig-dialog.component"
 
 export class LwClientSecurityConfigComponent implements OnInit, AfterViewInit {
     dataSource: MatTableDataSource<any>;
-    displayedColumns: string[] = ['Client Endpoint', 'Security Mode', 'Security Information','Action'];
+    displayedColumns: string[] = ['Client Endpoint', 'Security Mode', 'Security Information', 'Action'];
     @ViewChild(MatPaginator) paginator: MatPaginator;
-    clientsArray : any[];
+    clientsArray: any[];
     dataModelComponent: Boolean = false;
-    clientsCounter: number ;
+    clientsCounter: number;
     clientEndpoint: string;
-    constructor(private http: HttpClient,public dialog: MatDialog) { }
+    constructor(private http: HttpClient, public dialog: MatDialog, private lwService: LwService) { }
 
     ngAfterViewInit(): void {
         this.getClients();
-        
+
     }
     ngOnInit(): void {
-        
+
     }
 
-    openDialog(){
-        this.dialog.open(newConfigDialog,{
+    openDialog() {
+        this.dialog.open(newConfigDialog, {
             height: '450px',
             width: '240px',
         }).afterClosed().subscribe((clientsData) => {
@@ -39,8 +40,8 @@ export class LwClientSecurityConfigComponent implements OnInit, AfterViewInit {
         })
     }
 
-    getClients(){
-        
+    getClients() {
+
         this.http.get<any[]>('http://localhost:8080/api/v1/Lw/clientsSecurity', { withCredentials: true }).subscribe((clientsData) => {
             this.dataSource = new MatTableDataSource(clientsData)
             this.dataSource.paginator = this.paginator;
@@ -50,15 +51,19 @@ export class LwClientSecurityConfigComponent implements OnInit, AfterViewInit {
 
         })
     }
-   async deletClients(id){
-      await  this.http.delete('http://localhost:8080/api/v1/Lw/deleteClientsSecurity/?endpoint=' + id).toPromise().then((dta) => { })
-        this.getClients();
-        
+    async deletClients(id) {
+        let confirmation = confirm('Deleting client. Are you sure?');
+        if (confirmation == true) {
+            await this.http.delete('http://localhost:8080/api/v1/Lw/deleteClientsSecurity/?endpoint=' + id).toPromise().then((dta) => { })
+            this.getClients();
+            this.lwService.progress('DELETED', true);
+        }
+
     }
 
-    clientsSearch(event){
-        if (event.target.value == "") {this.dataSource.data = this.clientsArray;}
-        
+    clientsSearch(event) {
+        if (event.target.value == "") { this.dataSource.data = this.clientsArray; }
+
         else {
             let arrayContainer = [];
             this.clientsArray.forEach((element) => {
@@ -69,7 +74,7 @@ export class LwClientSecurityConfigComponent implements OnInit, AfterViewInit {
             this.dataSource.data = arrayContainer;
         }
     }
-    
+
 
 
 }
